Moving the CheckFeatPrerequisites into CraftingLogic
